// src/types/booking-output20240813.ts
export type BookingOutput20240813 = {
  absentHost: boolean;
  attendees: Attendee[];
  /**
   * Booking field responses consisting of an object with booking field slug as keys and user response as values.
   */
  bookingFieldsResponses: Record<string, any>;
  cancellationReason?: string | undefined;
  cancelledByEmail?: string | undefined;
  createdAt: string;
  description: string;
  duration: number;
  end: string;
  eventType: EventType;
  /**
   * Deprecated - rely on 'eventType' object containing the id instead.
   */
  eventTypeId: number;
  guests?: string[] | undefined;
  hosts: BookingHost[];
  id: number;
  location: string;
  /**
   * Deprecated - rely on 'location' field instead.
   */
  meetingUrl?: string | undefined;
  metadata?: Record<string, any> | undefined;
  rating?: number | undefined;
  rescheduledByEmail?: string | undefined;
  rescheduledFromUid?: string | undefined;
  reschedulingReason?: string | undefined;
  start: string;
  status: "accepted" | "cancelled" | "pending" | "rejected";
  title: string;
  uid: string;
  updatedAt: string;
};

/**
 * @internal
 * BookingOutput20240813 without any key transformation, this is what
 * we expect to come in as network data
 */
export type External$BookingOutput20240813 = {
  absentHost: boolean;
  attendees: External$Attendee[];
  bookingFieldsResponses: Record<string, any>;
  cancellationReason?: string | undefined;
  cancelledByEmail?: string | undefined;
  createdAt: string;
  description: string;
  duration: number;
  end: string;
  eventType: External$EventType;
  eventTypeId: number;
  guests?: string[] | undefined;
  hosts: External$BookingHost[];
  id: number;
  location: string;
  meetingUrl?: string | undefined;
  metadata?: Record<string, any> | undefined;
  rating?: number | undefined;
  rescheduledByEmail?: string | undefined;
  rescheduledFromUid?: string | undefined;
  reschedulingReason?: string | undefined;
  start: string;
  status: "accepted" | "cancelled" | "pending" | "rejected";
  title: string;
  uid: string;
  updatedAt: string;
};

/**
 * Takes network data, validates it, and transforms keys to match typescript object BookingOutput20240813
 */
const SchemaIn$BookingOutput20240813: z.ZodType<
  BookingOutput20240813, // output type of this zod object
  z.ZodTypeDef,
  unknown
> = z
  .object({
    absentHost: z.boolean(),
    attendees: z.array(Schemas$Attendee.in),
    bookingFieldsResponses: z.record(z.string(), z.any()),
    cancellationReason: z.string().optional(),
    cancelledByEmail: z.string().optional(),
    createdAt: z.string(),
    description: z.string(),
    duration: z.number(),
    end: z.string(),
    eventType: Schemas$EventType.in,
    eventTypeId: z.number(),
    guests: z.array(z.string()).optional(),
    hosts: z.array(Schemas$BookingHost.in),
    id: z.number(),
    location: z.string(),
    meetingUrl: z.string().optional(),
    metadata: z.record(z.string(), z.any()).optional(),
    rating: z.number().optional(),
    rescheduledByEmail: z.string().optional(),
    rescheduledFromUid: z.string().optional(),
    reschedulingReason: z.string().optional(),
    start: z.string(),
    status: z.enum(["accepted", "cancelled", "pending", "rejected"]),
    title: z.string(),
    uid: z.string(),
    updatedAt: z.string(),
  })
  .transform((obj) => {
    return zodTransform(obj, {
      absentHost: "absentHost",
      attendees: "attendees",
      bookingFieldsResponses: "bookingFieldsResponses",
      cancellationReason: "cancellationReason",
      cancelledByEmail: "cancelledByEmail",
      createdAt: "createdAt",
      description: "description",
      duration: "duration",
      end: "end",
      eventType: "eventType",
      eventTypeId: "eventTypeId",
      guests: "guests",
      hosts: "hosts",
      id: "id",
      location: "location",
      meetingUrl: "meetingUrl",
      metadata: "metadata",
      rating: "rating",
      rescheduledByEmail: "rescheduledByEmail",
      rescheduledFromUid: "rescheduledFromUid",
      reschedulingReason: "reschedulingReason",
      start: "start",
      status: "status",
      title: "title",
      uid: "uid",
      updatedAt: "updatedAt",
    });
  });

/**
 * @internal
 * Takes typescript data, validates it, and maps keys to match the expected external object External$BookingOutput20240813
 */
const SchemaOut$BookingOutput20240813: z.ZodType<
  External$BookingOutput20240813, // output type of this zod object
  z.ZodTypeDef,
  BookingOutput20240813 // the object to be transformed
> = z
  .object({
    absentHost: z.boolean(),
    attendees: z.array(Schemas$Attendee.out),
    bookingFieldsResponses: z.record(z.string(), z.any()),
    cancellationReason: z.string().optional(),
    cancelledByEmail: z.string().optional(),
    createdAt: z.string(),
    description: z.string(),
    duration: z.number(),
    end: z.string(),
    eventType: Schemas$EventType.out,
    eventTypeId: z.number(),
    guests: z.array(z.string()).optional(),
    hosts: z.array(Schemas$BookingHost.out),
    id: z.number(),
    location: z.string(),
    meetingUrl: z.string().optional(),
    metadata: z.record(z.string(), z.any()).optional(),
    rating: z.number().optional(),
    rescheduledByEmail: z.string().optional(),
    rescheduledFromUid: z.string().optional(),
    reschedulingReason: z.string().optional(),
    start: z.string(),
    status: z.enum(["accepted", "cancelled", "pending", "rejected"]),
    title: z.string(),
    uid: z.string(),
    updatedAt: z.string(),
  })
  .transform((obj) => {
    return zodTransform(obj, {
      absentHost: "absentHost",
      attendees: "attendees",
      bookingFieldsResponses: "bookingFieldsResponses",
      cancellationReason: "cancellationReason",
      cancelledByEmail: "cancelledByEmail",
      createdAt: "createdAt",
      description: "description",
      duration: "duration",
      end: "end",
      eventType: "eventType",
      eventTypeId: "eventTypeId",
      guests: "guests",
      hosts: "hosts",
      id: "id",
      location: "location",
      meetingUrl: "meetingUrl",
      metadata: "metadata",
      rating: "rating",
      rescheduledByEmail: "rescheduledByEmail",
      rescheduledFromUid: "rescheduledFromUid",
      reschedulingReason: "reschedulingReason",
      start: "start",
      status: "status",
      title: "title",
      uid: "uid",
      updatedAt: "updatedAt",
    });
  });

export const Schemas$BookingOutput20240813 = {
  in: SchemaIn$BookingOutput20240813,
  out: SchemaOut$BookingOutput20240813,
};